draft version of the STO-1G calculation engine
A rough first attempt at the actual quantum chemistry engine that will
make this work.  The demo.html page shows how to create inputs and
outputs.  The fchk file should be passable to jsmol (it is one of the
standard formats it reads…might be under “Formatted Gaussian Checkpoint
File”)